+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Implemented skipping validation/model if it is deprecated
 - Added kotsu exception DeprecatedEntityError, and use instead of ValueError for deprecated entity
 
 ### Changed